+++ CHANGELOG.rst
@@ -40,6 +40,8 @@ unreleased
 - Add migration for django_comments
 - Upgrade Django to 1.8.14
 - Upgrade to django-tinymce 2.4.0 (Mr. Senko)
+- Refactor tcms.core.migrations.0001_django_comments__object_pk.
+  PR #157 (Mr. Senko)
 
 3.8.18.04 (Jan 24 2017, released on MrSenko.com)
 ------------------------------------------------